requirement.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-rest-passwordreset==1.4.1
djangorestframework==3.15.1
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
